provide option for male/female of fittings on all connectors
The option can be selected separately for the east, north, west, south
fitting, with the compass direction indicating what's seen in OpenSCAD
rendering (at least for those of us in the northern hemisphere).

The male/female option is ignored for the zero-length connector
because there would never be room. For all other connectors, it's
up to you to make sure there is room for any female fittings.

The male/female option is not used for any of the dividers. A straight
connector with two female fittings is the same as a straight divider
of equivalent length.